Rename handlers as action files
rename files to match the functionality
