+++ common/models/workspace.js
@@ -7,8 +7,6 @@
 const async = require('async');
 const templateRegistry = require('../../lib/template-registry');
 const WorkspaceManager = require('../../lib/workspace-manager');
-const WorkspaceHandler = require('../../lib/workspace-handler');
-const TemplateHandler = require('../../lib/template-handler');
 const loopback = require('loopback');
 const boot = require('loopback-boot');
 const path = require('path');
