+++ lib/actions/datasource.js
@@ -1,8 +1,8 @@
 'use strict';
 
-const DataSource = require('./datamodel/datasource');
-const fsUtility = require('./util/file-utility');
-const mixin = require('./util/mixin');
+const DataSource = require('../datamodel/datasource');
+const fsUtility = require('../util/file-utility');
+const mixin = require('../util/mixin');
 
 class DataSourceActions {
   create(cb) {
