+++ lib/actions/facet.js
@@ -1,9 +1,9 @@
 'use strict';
 
-const config = require('./config');
-const Facet = require('./datamodel/facet');
-const fsUtility = require('./util/file-utility');
-const mixin = require('./util/mixin');
+const config = require('../config');
+const Facet = require('../datamodel/facet');
+const fsUtility = require('../util/file-utility');
+const mixin = require('../util/mixin');
 
 class FacetAction {
   refresh(cb) {
