+++ lib/actions/load.js
@@ -1,18 +1,18 @@
 'use strict';
 
-const DataSource = require('./datamodel/datasource');
-const Facet = require('./datamodel/facet');
-const fsUtility = require('./util/file-utility');
+const DataSource = require('../datamodel/datasource');
+const Facet = require('../datamodel/facet');
+const fsUtility = require('../util/file-utility');
 const lodash = require('lodash');
-const mixin = require('./util/mixin');
-const Model = require('./datamodel/model');
-const ModelConfig = require('./datamodel/model-config');
-const ModelMethod = require('./datamodel/model-method');
-const ModelProperty = require('./datamodel/model-property');
-const MiddlewarePhase = require('./datamodel/middleware-phase');
-const PackageDefinition = require('./datamodel/package-definition');
+const mixin = require('../util/mixin');
+const Model = require('../datamodel/model');
+const ModelConfig = require('../datamodel/model-config');
+const ModelMethod = require('../datamodel/model-method');
+const ModelProperty = require('../datamodel/model-property');
+const MiddlewarePhase = require('../datamodel/middleware-phase');
+const PackageDefinition = require('../datamodel/package-definition');
 const path = require('path');
-const Workspace = require('./workspace');
+const Workspace = require('../workspace');
 
 class WorkspaceActions {
   loadAll(cb) {
