+++ lib/actions/middleware.js
@@ -1,9 +1,9 @@
 'use strict';
 
-const fsUtility = require('./util/file-utility');
-const mixin = require('./util/mixin');
-const Middleware = require('./datamodel/middleware');
-const MiddlewarePhase = require('./datamodel/middleware-phase');
+const fsUtility = require('../util/file-utility');
+const mixin = require('../util/mixin');
+const Middleware = require('../datamodel/middleware');
+const MiddlewarePhase = require('../datamodel/middleware-phase');
 
 class MiddlewareActions {
   create(phaseName, cb) {
