+++ lib/actions/model-config.js
@@ -1,12 +1,8 @@
 'use strict';
 
-const workspaceHandler = require('./workspace-handler');
-const Model = require('./datamodel/model');
-const ModelConfig = require('./datamodel/model-config');
-const ModelProperty = require('./datamodel/model-property');
-const ModelMethod = require('./datamodel/model-method');
-const mixin = require('./util/mixin');
-const fsUtility = require('./util/file-utility');
+const ModelConfig = require('../datamodel/model-config');
+const mixin = require('../util/mixin');
+const fsUtility = require('../util/file-utility');
 
 class ModelConfigActions {
   create(modelId, facetName, cb) {
