+++ lib/actions/model-property.js
@@ -1,12 +1,9 @@
 'use strict';
 
-const workspaceHandler = require('./workspace-handler');
-const Model = require('./datamodel/model');
-const ModelConfig = require('./datamodel/model-config');
-const ModelProperty = require('./datamodel/model-property');
-const ModelMethod = require('./datamodel/model-method');
-const mixin = require('./util/mixin');
-const fsUtility = require('./util/file-utility');
+const ModelProperty = require('../datamodel/model-property');
+const ModelMethod = require('../datamodel/model-method');
+const mixin = require('../util/mixin');
+const fsUtility = require('../util/file-utility');
 
 class ModelPropertyActions {
   create(modelId, cb) {
