+++ lib/actions/model.js
@@ -1,11 +1,11 @@
 'use strict';
 
-const Model = require('./datamodel/model');
-const ModelConfig = require('./datamodel/model-config');
-const ModelProperty = require('./datamodel/model-property');
-const ModelMethod = require('./datamodel/model-method');
-const mixin = require('./util/mixin');
-const fsUtility = require('./util/file-utility');
+const Model = require('../datamodel/model');
+const ModelConfig = require('../datamodel/model-config');
+const ModelProperty = require('../datamodel/model-property');
+const ModelMethod = require('../datamodel/model-method');
+const mixin = require('../util/mixin');
+const fsUtility = require('../util/file-utility');
 
 class ModelActions {
   create(cb) {
