+++ lib/actions/template.js
@@ -1,14 +1,14 @@
 'use strict';
 
 const async = require('async');
-const DataSource = require('./datamodel/datasource');
-const Facet = require('./datamodel/facet');
+const DataSource = require('../datamodel/datasource');
+const Facet = require('../datamodel/facet');
 const fs = require('fs-extra');
-const fsUtility = require('./util/file-utility');
-const mixin = require('./util/mixin');
-const ModelConfig = require('./datamodel/model-config');
-const Middleware = require('./datamodel/middleware');
-const Workspace = require('./workspace');
+const fsUtility = require('../util/file-utility');
+const mixin = require('../util/mixin');
+const ModelConfig = require('../datamodel/model-config');
+const Middleware = require('../datamodel/middleware');
+const Workspace = require('../workspace');
 const path = require('path');
 
 class TemplateActions {
@@ -99,12 +99,12 @@ function getTemplateFiles(template) {
   if (template.dirList) {
     template.dirList.forEach(function(dir) {
       let filePath = path.join(__dirname,
-        '../templates/files', dir);
+        '../../templates/files', dir);
       templateFiles.push(filePath);
     });
   }
   let filePath = path.join(__dirname,
-    '../templates/files',
+    '../../templates/files',
     template.files.path);
   templateFiles.push(filePath);
   return templateFiles;
