+++ lib/workspace-manager.js
@@ -2,13 +2,14 @@
 const config = require('./config.json');
 const Workspace = require('./workspace');
 const templateRegistry = require('./template-registry');
-const dataSourceHandler = require('./data-source-handler');
-const facetHandler = require('./facet-handler');
-const middlewareHandler = require('./middleware-handler');
-const modelHandler = require('./model-handler');
-const modelConfigHandler = require('./model-config-handler');
-const modelPropertyHandler = require('./model-property-handler');
-const workspaceHandler = require('./workspace-handler');
+const dataSourceActions = require('./actions/datasource');
+const facetActions = require('./actions/facet');
+const loadActions = require('./actions/load');
+const middlewareActions = require('./actions/middleware');
+const modelActions = require('./actions/model');
+const modelConfigActions = require('./actions/model-config');
+const modelPropertyActions = require('./actions/model-property');
+const templateActions = require('./actions/template');
 
 /**
  * @class WorkspaceManager
